pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lingye-agent"
version = "0.1.0"
description = "模块化 LLM Agent 框架：多范式 Agent + 多类型记忆 + RAG + MCP + GSSC 上下文工程"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{ name = "Lingye" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "openai>=1.30.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "tiktoken>=0.9.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
rag = [
    "qdrant-client>=1.6.0",
    "sentence-transformers>=2.5.0",
    "scikit-learn>=1.3.0",
]
graph = ["neo4j>=5.0.0"]
mcp = ["fastmcp>=2.0.0"]
nlp = ["jieba>=0.42.1", "dashscope>=1.14.0"]
doc = ["markitdown[pdf]>=0.0.1a3"]
ui = ["gradio>=4.0.0"]
all = ["lingye-agent[rag,graph,mcp,nlp,doc,ui]"]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0"]

[project.urls]
Homepage = "https://github.com/yourname/lingye-agent"
Repository = "https://github.com/yourname/lingye-agent"

[tool.hatch.build.targets.wheel]
packages = ["lingye_agent"]
